Move ComponentKey type into core.
